requirements.txt
grpcio==1.60.1
google-generativeai